steer_image/requirements.txt
# STEER Image Retrieval 额外依赖
open_clip_torch>=2.20.0
faiss-cpu>=1.7.4
torchvision>=0.11.0
numpy>=1.21.0
pyyaml>=6.0
tqdm>=4.62.0
Pillow>=9.0.0
